test: refactor dos testes de user e criação do arquivo para criar  o userDTO e userObj
